refactor: consolidate checkpointer functionality into common module
- Moved checkpointer logic from the Kronos agent to a shared CommonCheckpointer module for better reusability and maintainability.
- Updated the KronosAgentBuilder to utilize the new CommonCheckpointer, simplifying the initialization process.
- Removed redundant Kronos-specific checkpointer configuration and validation code, streamlining the overall architecture.
