+++ apps/server/.env.example
@@ -1,11 +1,15 @@
 NODE_ENV=
+
+DATABASE_URI=
 DATABASE_HOST=
 DATABASE_PORT=
 DATABASE_USERNAME=
 DATABASE_PASSWORD=
 DATABASE_NAME=
 DATABASE_LOGGING=
+
 COMPOSIO_API_KEY=
+
 LANGSMITH_TRACING=
 LANGSMITH_API_KEY=
 LANGSMITH_PROJECT=
